pybot/requirements.txt
python-telegram-bot==20.7
influxdb-client==1.47.0
pytz==2023.3.post1
python-dotenv==1.0.0
groq==0.12.0
anthropic==0.7.0
